Add HTTPS examples and Certificates.json manifest file
This forms part of 5G-MAG/rt-5gms-application-server#9.

The Certificates.json file is a simple manifest file which maps certificate Ids to the filename of the PEM file which holds the public certificate and private key. Relative filenames in the file will be treated as being relative to the Certificates.json file itself.

The make_self_signed_certs.py Python script will take a ContentHostingConfiguration and Certificates JSON files and will create a self-signed certificate PEM file using the canonicalDomainName and domainNameAlias from the ContentHostingConfiguration to fill in the Subject Common Name and subjectAltName extensions, and place the resulting certificate in the file indicated by the Certificates JSON file. It will only create certificates for those distribution points mentioned in the ContentHostingConfiguration.

The ContentHostingConfiguration_Big-Buck-Bunny_pull-ingest_https.json file is an example ContentHostingConfiguration to create an HTTPS distribution point instead of HTTP.

The ContentHostingConfiguration_Big-Buck-Bunny_pull-ingest_http_and_https.json file is an example ContentHostingConfiguration to create distribution points for both HTTP and HTTPS from the same pull ingest.