Enhancements to OpenAPI generator script.
Support for 5G-MAG/rt-5gms-application-function#2.

Change -a option to allow bindings for multiple APIs to be generated at once.

Add -y option to take an executable path of a process to run after cloning the 5G_APIs repository, but before generating the bindings. This allows the rt-5gms-application-function to "fix" the YAML so that openapi-generator v5.2.0 can generate correct bindings. Version 5.2.0 cannot handle an anyOf being used to allow future expansion of an enum by having two string type schemas in the anyOf, one to hanbdle the enum and another to handle any other string in the field.