Change version badge to latest release instead of latest tag and add links
Changed the version badge to a fixed "No release yet" message and disabled the auto-generated link. Left commented tags to use the release badge and latest release link for the version badge for future use.

Added a link to the License badge to take you to the 5G-MAG license PDF.

Removed the auto-generated link from the Status badge.